+++ app/lib/features/main/presentation/pages/_pages.dart
@@ -1,25 +1,18 @@
-import 'dart:io';
-
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:go_router/go_router.dart';
 import 'package:google_sign_in/google_sign_in.dart';
-import 'package:intl/intl.dart';
 import 'package:url_launcher/url_launcher.dart';
-import 'package:yaml/yaml.dart';
 
-import '../../../../core/bases/enum/button_type.dart';
 import '../../../../core/bases/widgets/atoms/button.dart';
 import '../../../../core/bases/widgets/layout/bebras_scaffold.dart';
 import '../../../../core/bases/widgets/layout/bottom_navbar.dart';
 import '../../../../core/constants/assets.dart';
 import '../../../../core/theme/font_theme.dart';
 import '../../../../services/firebase_service.dart';
-import '../bloc/home_cubit.dart';
 
 part 'home_page.dart';
+part 'policy_page.dart';
 part 'setting_page.dart';
