+++ app/lib/features/main/presentation/pages/clip_path.dart
