+++ app/lib/features/main/presentation/pages/home_page.dart
@@ -1,157 +1,169 @@
 part of '_pages.dart';
 
-class HomePage extends StatefulWidget {
-  const HomePage({super.key});
+class HomePageV2 extends StatefulWidget {
+  const HomePageV2({super.key});
 
   @override
-  State<HomePage> createState() => _HomePageState();
+  State<HomePageV2> createState() => _HomePageV2State();
 }
 
-class Course {
-  String title;
-  String description;
-
-  Course(this.title, this.description);
-}
-
-class _HomePageState extends State<HomePage> {
+class _HomePageV2State extends State<HomePageV2> {
   String version = '';
 
   @override
   void initState() {
-    context.read<HomeCubit>().fetchUser();
-    _loadVersion();
     super.initState();
   }
 
-  Future<void> _loadVersion() async {
-    try {
-      final yamlMap =
-          loadYaml(await rootBundle.loadString('pubspec.yaml')); 
-      final versionWithBuildNumber =
-          yamlMap['version'] as String; 
-      final parts = versionWithBuildNumber.split('+'); 
-      setState(() {
-        version = parts[0];
-      });
-    } catch (e) {
-      print('Error loading version: $e');
-    }
+  Future<void> showModal() async {
+    return showModalBottomSheet<void>(
+      context: context,
+      builder: (BuildContext context) {
+        return StatefulBuilder(
+          builder: (BuildContext context, StateSetter setState) {
+            return Container(
+              height: 260,
+              width: double.infinity,
+              color: Colors.white,
+              child: Column(
+                mainAxisSize: MainAxisSize.min,
+                children: <Widget>[
+                  const SizedBox(
+                    height: 40,
+                  ),
+                  Container(
+                    margin: const EdgeInsets.only(left: 20),
+                    child: const Text(
+                      'Ayo pilih latihanmu!',
+                      textAlign: TextAlign.left,
+                      style:
+                          TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
+                    ),
+                  ),
+                  const SizedBox(
+                    height: 25,
+                  ),
+                  Container(
+                    padding: const EdgeInsets.symmetric(horizontal: 40),
+                    width: double.infinity,
+                    child: const Button(
+                      customButtonColor: const Color(0xFF1BB8E1),
+                      customTextColor: Colors.white,
+                      text: 'Unduh Latihan Minggu Ini',
+                    ),
+                  ),
+                  const SizedBox(
+                    height: 20,
+                  ),
+                  Container(
+                    padding: const EdgeInsets.symmetric(horizontal: 40),
+                    width: double.infinity,
+                    child: const Button(
+                      customTextColor: const Color(0xFF1BB8E1),
+                      customButtonColor: const Color(0x1F1BB8E1),
+                      text: 'Latihan Minggu Depan',
+                    ),
+                  ),
+                ],
+              ),
+            );
+          },
+        );
+      },
+    ).whenComplete(() => null);
   }
 
   @override
   Widget build(BuildContext context) {
     return BebrasScaffold(
-      body: SingleChildScrollView(
+      body: Container(
+        color: const Color(0xFF1BB8E1),
         child: Stack(
           children: [
-            Padding(
-              padding: const EdgeInsets.all(32),
-              child: Column(
-                children: [
-                  Image.asset(
-                    Assets.bebrasPandaiText,
-                  ),
-                  const SizedBox(
-                    height: 40,
-                  ),
-                  BlocBuilder<HomeCubit, HomeState>(builder: (context, state) {
-                    if (state is HomeSuccess) {
-                      return RichText(
-                        text: TextSpan(
-                          text: 'Selamat Datang\n',
-                          style: FontTheme.blackTitle(),
-                          children: <TextSpan>[
-                            TextSpan(
-                                text: toBeginningOfSentenceCase(
-                                  '${state.user.name}!',
-                                ),
-                                style: FontTheme.blackTitleBold()),
-                          ],
-                        ),
-                      );
-                    }
-                    return Container();
-                  }),
-                  const SizedBox(
-                    height: 30,
-                  ),
-                  Button(
-                    onTap: () async {
-                      final url = Uri.parse(
-                        'https://bebras.or.id/v3/bebras-indonesia-challenge-2023/',
-                      );
-                      if (!await launchUrl(url)) {
-                        throw Exception('Could not launch $url');
-                      }
-                    },
-                    customButtonColor: Colors.blue,
-                    customTextColor: Colors.white,
-                    text: 'ℹ️  Tentang Tantangan Bebras  ℹ️',
-                  ),
-                  const SizedBox(
-                    height: 25,
-                  ),
-                  Button(
-                    buttonType: ButtonType.primary,
-                    onTap: () async {
-                      await context.push('/material');
-                    },
-                    text: 'Lihat / Cetak Materi',
-                  ),
-                  const SizedBox(
-                    height: 25,
-                  ),
-                  Button(
-                    buttonType: ButtonType.primary,
-                    onTap: () async {
-                      await context.push('/quiz_registration');
-                    },
-                    text: 'Ikut Latihan Mingguan',
-                  ),
-                  const SizedBox(
-                    height: 25,
-                  ),
-                  Button(
-                    onTap: () async {
-                      await context.push('/setting');
-                    },
-                    customButtonColor: Colors.grey,
-                    customTextColor: Colors.white,
-                    text: 'Pengaturan',
-                  ),
-                  SizedBox(
-                    height: MediaQuery.of(context).size.height - 640,
-                  ),
-                  InkWell(
-                    onTap: () async {
-                      final url = Uri.parse(
-                        'https://tlx.toki.id/',
-                      );
-                      if (!await launchUrl(url)) {
-                        throw Exception('Could not launch $url');
-                      }
-                    },
-                    child: Center(
-                      child: Text(
-                        'From Ikatan Alumni TOKI with ❤️',
-                        textAlign: TextAlign.center,
-                        style: FontTheme.greyNormal14(),
-                      ),
+            Column(
+              children: [
+                const SizedBox(
+                  height: 60,
+                ),
+                Container(
+                  alignment: Alignment.center,
+                  child: ClipRRect(
+                    borderRadius: BorderRadius.circular(24),
+                    child: Image.asset(
+                      Assets.bLogo,
+                      height: 150,
                     ),
                   ),
-                  Center(
-                      child: Text(
-                        'V $version',
-                        textAlign: TextAlign.center,
-                        style: FontTheme.greyNormal14(),
-                      ),
+                ),
+              ],
+            ),
+            Align(
+              alignment: Alignment.bottomCenter,
+              child: Container(
+                height: 300,
+                width: double.infinity,
+                decoration: const BoxDecoration(
+                  color: Colors.white,
+                  borderRadius: BorderRadius.only(
+                    topLeft: Radius.circular(42),
+                    topRight: Radius.circular(42),
                   ),
-                ],
+                ),
+                child: Column(
+                  mainAxisAlignment: MainAxisAlignment.center,
+                  children: [
+                    const Text(
+                      'Ayo mulai latihanmu!',
+                      style: TextStyle(
+                        fontWeight: FontWeight.w600,
+                        fontSize: 18,
+                      ),
+                    ),
+                    const SizedBox(
+                      height: 30,
+                    ),
+                    SizedBox(
+                      width: 230,
+                      child: Button(
+                        customTextColor: Colors.white,
+                        customButtonColor: Color(0xFF1BB8E1),
+                        fontSize: 14,
+                        innerVerticalPadding: 14,
+                        onTap: () async {
+                          await showModal();
+                        },
+                        text: 'Daftar Latihan',
+                      ),
+                    ),
+                    const SizedBox(
+                      height: 10,
+                    ),
+                    const SizedBox(
+                      width: 230,
+                      child: Button(
+                        customTextColor: const Color(0xFF1BB8E1),
+                        customButtonColor: const Color(0x1F1BB8E1),
+                        fontSize: 14,
+                        innerVerticalPadding: 14,
+                        text: 'Riwayat Latihan',
+                      ),
+                    ),
+                  ],
+                ),
               ),
             ),
           ],
         ),
+      ),
+      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          context.go('/quiz_registration');
+        },
+        child: const FaIcon(FontAwesomeIcons.graduationCap),
+      ),
+      bottomNavigationBar: const BottomNavBar(
+        currentIndex: 2,
       ),
     );
   }
