--- app/lib/features/main/presentation/pages/v2/home_page.dart
@@ -1,170 +1,0 @@
-part of '_pages.dart';
-
-class HomePageV2 extends StatefulWidget {
-  const HomePageV2({super.key});
-
-  @override
-  State<HomePageV2> createState() => _HomePageV2State();
-}
-
-class _HomePageV2State extends State<HomePageV2> {
-  String version = '';
-
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  Future<void> showModal() async {
-    return showModalBottomSheet<void>(
-      context: context,
-      builder: (BuildContext context) {
-        return StatefulBuilder(
-          builder: (BuildContext context, StateSetter setState) {
-            return Container(
-              height: 260,
-              width: double.infinity,
-              color: Colors.white,
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: <Widget>[
-                  const SizedBox(
-                    height: 40,
-                  ),
-                  Container(
-                    margin: const EdgeInsets.only(left: 20),
-                    child: const Text(
-                      'Ayo pilih latihanmu!',
-                      textAlign: TextAlign.left,
-                      style:
-                          TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-                    ),
-                  ),
-                  const SizedBox(
-                    height: 25,
-                  ),
-                  Container(
-                    padding: const EdgeInsets.symmetric(horizontal: 40),
-                    width: double.infinity,
-                    child: const Button(
-                      customButtonColor: const Color(0xFF1BB8E1),
-                      customTextColor: Colors.white,
-                      text: 'Unduh Latihan Minggu Ini',
-                    ),
-                  ),
-                  const SizedBox(
-                    height: 20,
-                  ),
-                  Container(
-                    padding: const EdgeInsets.symmetric(horizontal: 40),
-                    width: double.infinity,
-                    child: const Button(
-                      customTextColor: const Color(0xFF1BB8E1),
-                      customButtonColor: const Color(0x1F1BB8E1),
-                      text: 'Latihan Minggu Depan',
-                    ),
-                  ),
-                ],
-              ),
-            );
-          },
-        );
-      },
-    ).whenComplete(() => null);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return BebrasScaffold(
-      body: Container(
-        color: const Color(0xFF1BB8E1),
-        child: Stack(
-          children: [
-            Column(
-              children: [
-                const SizedBox(
-                  height: 60,
-                ),
-                Container(
-                  alignment: Alignment.center,
-                  child: ClipRRect(
-                    borderRadius: BorderRadius.circular(24),
-                    child: Image.asset(
-                      Assets.bLogo,
-                      height: 150,
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            Align(
-              alignment: Alignment.bottomCenter,
-              child: Container(
-                height: 300,
-                width: double.infinity,
-                decoration: const BoxDecoration(
-                  color: Colors.white,
-                  borderRadius: BorderRadius.only(
-                    topLeft: Radius.circular(42),
-                    topRight: Radius.circular(42),
-                  ),
-                ),
-                child: Column(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    const Text(
-                      'Ayo mulai latihanmu!',
-                      style: TextStyle(
-                        fontWeight: FontWeight.w600,
-                        fontSize: 18,
-                      ),
-                    ),
-                    const SizedBox(
-                      height: 30,
-                    ),
-                    SizedBox(
-                      width: 230,
-                      child: Button(
-                        customTextColor: Colors.white,
-                        customButtonColor: Color(0xFF1BB8E1),
-                        fontSize: 14,
-                        innerVerticalPadding: 14,
-                        onTap: () async {
-                          await showModal();
-                        },
-                        text: 'Daftar Latihan',
-                      ),
-                    ),
-                    const SizedBox(
-                      height: 10,
-                    ),
-                    const SizedBox(
-                      width: 230,
-                      child: Button(
-                        customTextColor: const Color(0xFF1BB8E1),
-                        customButtonColor: const Color(0x1F1BB8E1),
-                        fontSize: 14,
-                        innerVerticalPadding: 14,
-                        text: 'Riwayat Latihan',
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          context.go('/quiz_registration');
-        },
-        child: const FaIcon(FontAwesomeIcons.graduationCap),
-      ),
-      bottomNavigationBar: const BottomNavBar(
-        currentIndex: 2,
-      ),
-    );
-  }
-}
