+++ app/lib/features/material/menu/presentation/pages/_pages.dart
@@ -2,12 +2,12 @@ import 'dart:io';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
+import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:go_router/go_router.dart';
 import 'package:printing/printing.dart';
 
-import '../../../../../core/bases/enum/button_type.dart';
-import '../../../../../core/bases/widgets/atoms/button.dart';
 import '../../../../../core/bases/widgets/layout/bebras_scaffold.dart';
+import '../../../../../core/bases/widgets/layout/bottom_navbar.dart';
 import '../../../../../core/constants/assets.dart';
 import '../../data/repositories/material.dart';
 import '../model/bebras_group_category.dart';
